예약 상세페이지에서 리뷰 작성 팝업

+++ lib/data/models/review.dart
@@ -1,5 +1,5 @@
 class Review {
-  final double rating; // 별점
+  final int rating; // 별점
   final String comment; // 코멘트
 
   Review({
